% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_grad}
\alias{compute_grad}
\title{Compute Gradient (Internal C++ function)}
\usage{
compute_grad(X, D_x, a, b, eps = 1e-20)
}
\arguments{
\item{X}{Numeric Matrix}

\item{D_x}{Numeric Matrix}

\item{a}{Numeric Matrix}

\item{b}{Numeric Matrix}

\item{eps}{Double}
}
\value{
Gradient matrix
}
\description{
Compute Gradient (Internal C++ function)
}
