% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_DOST_spatial}
\alias{plot_DOST_spatial}
\title{Plot DOST Spatial Clusters}
\usage{
plot_DOST_spatial(coords, labels, point_size = 1)
}
\arguments{
\item{coords}{A data frame or matrix of spatial coordinates (Spots x 2).}

\item{labels}{A vector of cluster labels.}

\item{point_size}{Numeric, size of the points (default 1).}
}
\value{
A ggplot object.
}
\description{
Visualizes the clustering results in spatial coordinates.
}
