% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_DOST_umap}
\alias{plot_DOST_umap}
\title{Plot DOST Embedding (UMAP)}
\usage{
plot_DOST_umap(Z, labels)
}
\arguments{
\item{Z}{The embedding matrix from DOST.}

\item{labels}{A vector of cluster labels.}
}
\description{
Projects the DOST embedding into 2D using UMAP.
}
