% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOST-package.R
\docType{package}
\name{DOST-package}
\alias{DOST-package}
\title{DOST: Distance-preserving Optimization for Spatial Transcriptomics}
\description{
DOST is a distance-preserving optimization based method that clusters spatial transcriptomics (ST) spots into spatial domains. DOST embeds ST spots into a low-dimensional space using a Sammon mapping inspired loss function that jointly preserves gene expression similarity and spatial smoothness.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/misratasci/DOST}
  \item \url{https://misratasci.github.io/DOST/}
  \item Report bugs at \url{https://github.com/misratasci/DOST/issues}
}

}
\author{
\strong{Maintainer}: Mısra Taşçı \email{mtasci18@ku.edu.tr}

Authors:
\itemize{
  \item Mehmet Gönen \email{mehmetgonen@ku.edu.tr}
}

}
\keyword{internal}
